test: more tagging
