+++ deployment/ccip/changeset/aptos/cs_deploy_aptos_chain_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && aptos
+
 package aptos
 
 import (
