+++ deployment/ccip/changeset/aptos/cs_set_ocr3_offramp_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && aptos
+
 package aptos_test
 
 import (
