+++ deployment/ccip/changeset/ccip-attestation/cs_configure_evm_signer_registry_unit_test.go
@@ -1,3 +1,5 @@
+//go:build !integration
+
 package ccip_attestation_test
 
 import (
