+++ deployment/ccip/changeset/solana_v0_1_1/cs_solana_token_unit_test.go
@@ -1,3 +1,5 @@
+//go:build !integration
+
 package solana_test
 
 import (
