+++ deployment/ccip/changeset/solana_v0_1_1/transfer_ccip_to_mcms_with_timelock_unit_test.go
@@ -1,3 +1,5 @@
+//go:build !integration
+
 package solana_test
 
 import (
