+++ deployment/ccip/changeset/v1_5_1/cs_add_token_e2e_integ_test.go
@@ -1,3 +1,7 @@
+//go:build integration && !solana && !aptos
+
+// maybe?
+
 package v1_5_1_test
 
 import (
